+++ app/components/WalletContextProvider.tsx
@@ -1,0 +1,42 @@
+"use client";
+
+import {
+  WalletProvider,
+  ConnectionProvider,
+} from "@solana/wallet-adapter-react";
+import { WalletModalProvider } from "@solana/wallet-adapter-react-ui";
+import {
+  SolflareWalletAdapter,
+  AlphaWalletAdapter,
+  LedgerWalletAdapter,
+} from "@solana/wallet-adapter-wallets";
+import { FC, useMemo } from "react";
+
+import "@solana/wallet-adapter-react-ui/styles.css";
+import { clusterApiUrl } from "@solana/web3.js";
+
+type Props = {
+  children?: React.ReactNode;
+};
+
+export const WalletContextProvider: FC<Props> = ({ children }) => {
+  //input your RPC as your endpoint value
+  const endpoint = clusterApiUrl("devnet");
+
+  const wallets = useMemo(
+    () => [
+      new SolflareWalletAdapter(),
+      new AlphaWalletAdapter(),
+      new LedgerWalletAdapter(),
+    ],
+    []
+  );
+
+  return (
+    <ConnectionProvider endpoint={endpoint}>
+      <WalletProvider wallets={wallets}>
+        <WalletModalProvider>{children}</WalletModalProvider>
+      </WalletProvider>
+    </ConnectionProvider>
+  );
+};
