+++ app/layout.tsx
@@ -1,0 +1,27 @@
+import type { Metadata } from "next";
+import "./globals.css";
+import styles from "./styles/Home.module.css";
+import { WalletContextProvider } from "./components/WalletContextProvider";
+import { AppBar } from "./components/AppBar";
+
+export const metadata: Metadata = {
+  title: "Solana Token Mint",
+  description: "Generated by create next app",
+};
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <html lang="en">
+      <body className={styles.App}>
+        <WalletContextProvider>
+          <AppBar />
+          {children}
+        </WalletContextProvider>
+      </body>
+    </html>
+  );
+}
