+++ package.json
@@ -9,21 +9,23 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@solana/spl-token": "^0.2.0",
-    "@solana/wallet-adapter-base": "^0.9.5",
-    "@solana/wallet-adapter-react": "^0.15.4",
-    "@solana/wallet-adapter-react-ui": "^0.9.6",
-    "@solana/wallet-adapter-wallets": "^0.15.5",
-    "@solana/web3.js": "^1.37.0",
-    "next": "12.1.2",
-    "react": "18.0.0",
-    "react-dom": "18.0.0"
+    "@solana-developers/helpers": "^2.4.0",
+    "@solana/spl-token": "^0.4.8",
+    "@solana/wallet-adapter-base": "^0.9.23",
+    "@solana/wallet-adapter-react": "^0.15.35",
+    "@solana/wallet-adapter-react-ui": "^0.9.35",
+    "@solana/wallet-adapter-wallets": "^0.19.32",
+    "@solana/web3.js": "^1.91.7",
+    "next": "14.2.3",
+    "react": "^18.3.1",
+    "react-dom": "^18"
   },
   "devDependencies": {
-    "@types/node": "^17.0.23",
-    "@types/react": "^17.0.43",
-    "eslint": "8.12.0",
-    "eslint-config-next": "12.1.2",
-    "typescript": "^4.6.3"
+    "@types/node": "^20",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
+    "eslint": "^8",
+    "eslint-config-next": "14.2.3",
+    "typescript": "^5"
   }
 }
